Transplant blizzard rendering from BetterWeather
The rendering logic is presented as a static method
in `climate.BlizzardRenderer`.
Everything is static so there's no need to
manually initialize anything.

Now due to some conflicts with other part of the mod
(e.g. PrimalWinter, which also roadblock-injects
`net.minecraft.client.renderer.WorldRenderer.renderRainSnow`
to add its own snow rendering logic),
the integration of `BlizzardRenderer` into the game
is suspended. If you want to see it in action,
set a higher priority of `mixin.WorldRenderMixin`
and inject an always-blowing blizzard
to the game.
(Yeah, the capability for `climate.ClimateData`
may not be present as well,
due to some reason unknown to me.)

These arcane, dirty and evilly-written codes
are mainly borrowed from the mod BetterWeather.
But somehow it works, so let's forget about its
content.